+++ src/narcoticsense/spectroscopy/__init__.py
@@ -1,5 +1,6 @@
 from .core import SpectralDataset, Spectrum
 from .dataset import AlignedDataset, align_spectra, dataset_summary
+from .export import spectra_to_long_dataframe, write_spectra_long_csv
 from .importer import ImportResult, import_spectrum
 from .io import read_csv_spectrum
 
@@ -12,4 +13,6 @@ __all__ = [
     "ImportResult",
     "import_spectrum",
     "read_csv_spectrum",
+    "spectra_to_long_dataframe",
+    "write_spectra_long_csv",
 ]